+++ wallets/cosmostation-extension/types/cosmostation.d.ts
@@ -1,3 +1,2 @@
 import { CosmostationExtensionWallet } from './extension';
 export declare const wallets: CosmostationExtensionWallet[];
-export declare const walletNames: string[];