+++ wallets/cosmostation-mobile/src/cosmostation.ts
@@ -6,4 +6,3 @@ import {
 const cosmostationMobile = new CosmostationMobileWallet(cosmostationMobileInfo);
 
 export const wallets = [cosmostationMobile];
-export const walletNames = wallets.map((wallet) => wallet.walletName);