+++ wallets/leap-extension/src/leap.ts
@@ -3,4 +3,3 @@ import { leapExtensionInfo, LeapExtensionWallet } from './extension';
 const leapExtension = new LeapExtensionWallet(leapExtensionInfo);
 
 export const wallets = [leapExtension];
-export const walletNames = wallets.map((wallet) => wallet.walletName);