+++ wallets/leap/types/index.d.ts
@@ -1,3 +1,1 @@
-import { LeapExtensionWallet } from '@cosmos-kit/leap-extension';
-export declare const wallets: LeapExtensionWallet[];
-export declare const walletNames: string[];
+export declare const wallets: import("@cosmos-kit/leap-extension").LeapExtensionWallet[];