+++ wallets/vectis-extension/src/vectis.ts
@@ -3,4 +3,3 @@ import { vectisExtensionInfo, VectisExtensionWallet } from './extension';
 const vectisExtension = new VectisExtensionWallet(vectisExtensionInfo);
 
 export const wallets = [vectisExtension];
-export const walletNames = wallets.map((wallet) => wallet.walletName);