+++ wallets/vectis-extension/types/vectis.d.ts
@@ -1,3 +1,2 @@
 import { VectisExtensionWallet } from './extension';
 export declare const wallets: VectisExtensionWallet[];
-export declare const walletNames: string[];