+++ packages/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@1.0.2...@cosmos-kit/core@1.0.3) (2023-01-14)
+
+**Note:** Version bump only for package @cosmos-kit/core
+
+
+
+
+
 ## [1.0.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@1.0.1...@cosmos-kit/core@1.0.2) (2023-01-13)
 
 **Note:** Version bump only for package @cosmos-kit/core